+++ llm4pddl/envs/assets/pddl/dressed/generate_tasks.py
@@ -14,11 +14,11 @@ def _generate_dressing_problems(min_people: int, max_people: int,
                                 rng: np.random.Generator) -> List[str]:
     """Generates problems from a distribution.
 
-    max numbers are exclusive, so if you want exactly two people set min
+    Max numbers are exclusive, so if you want exactly two people set min
     to 2 and max to 3.
     """
     problems = []
-    for _ in range(1, num_probs + 1):
+    for _ in range(num_probs):
         num_people = rng.integers(min_people, max_people)
         num_casual = rng.integers(min_casual, max_casual)
         num_formal_dress = rng.integers(min_formal_dress, max_formal_dress)
@@ -38,7 +38,7 @@ def _generate_dressing_problem(num_people: int, num_casual_events: int,
                                num_formal_in_dress: int,
                                num_formal_in_suit: int, num_extra_objects: int,
                                rng: np.random.Generator) -> str:
-    """Generates a single problem with the given parmeters."""
+    """Generates a single problem with the given parameters."""
     assert num_people >= (num_casual_events + num_formal_in_dress +
                           num_formal_in_suit)
     # deciding extra clothing:
@@ -98,7 +98,7 @@ def _generate_dressing_problem(num_people: int, num_casual_events: int,
         if len(obj['objs']) > 0:
             object_str += ' '.join(
                 obj['objs']) + f' - {obj["type"]}\n            '
-    # making initial predicates:
+    # making initial ground atoms:
     init_preds = ''
     for person in persons:
         init_preds += (f'(wearing-nothing-formal {person})\n         ' +
@@ -106,7 +106,7 @@ def _generate_dressing_problem(num_people: int, num_casual_events: int,
     for clothing in (dresses + sweatpants + sweatshirts + nicepants +
                      collaredshirt + suitjacket):
         init_preds += f'(in-closet {clothing})' + '\n         '
-    # making goal predicates:
+    # making goal ground atoms:
     goals = ''
     goal_predicates = []
     rng.shuffle(persons)
@@ -131,67 +131,30 @@ def _generate_dressing_problem(num_people: int, num_casual_events: int,
     return problem_string
 
 
-def test_generate_dressing_problems() -> None:
-    """tests the function _generate_dressing_problems."""
-    rng = np.random.default_rng(seed=0)
-    prob1 = _generate_dressing_problems(2, 3, 2, 3, 0, 1, 0, 1, 0, 1, 1,
-                                        rng)[0]
-    rng2 = np.random.default_rng(seed=0)
-    reference_prob = _generate_dressing_problem(2, 2, 0, 0, 0, rng2)
-    assert prob1 == reference_prob
-
-
-def test_generate_dressing_problem() -> None:
-    """tests the function _generate_dressing_problem."""
-    rng = np.random.default_rng(seed=0)
-    out = _generate_dressing_problem(2, 2, 0, 0, 0, rng)
-    assert out == """(define (problem dressed)
-  (:domain dressed)
-  (:objects person1 person2 - person
-            sweatpants1 sweatpants2 - sweatpants
-            sweatshirt1 sweatshirt2 - sweatshirt
-            )
-  (:init (wearing-nothing-formal person1)
-         (wearing-nothing-casual person1)
-         (wearing-nothing-formal person2)
-         (wearing-nothing-casual person2)
-         (in-closet sweatpants1)
-         (in-closet sweatpants2)
-         (in-closet sweatshirt1)
-         (in-closet sweatshirt2)
-         )
-  (:goal (and (attending-casual-event person1)
-              (attending-casual-event person2)
-              ))
-  )"""
-
-
 if __name__ == "__main__":
     import os
     loc = os.path.dirname(os.path.realpath(__file__))
-    test_generate_dressing_problem()
-    test_generate_dressing_problems()
 
     # 4 levels of difficulty:
     rng_default = np.random.default_rng(seed=0)
-    thirty: List[str] = []
+    problems: List[str] = []
     # level one:
-    thirty += _generate_dressing_problems(4, 5, 0, 3, 0, 2, 0, 2, 0, 2, 7,
-                                          rng_default)
+    problems += _generate_dressing_problems(4, 5, 0, 3, 0, 2, 0, 2, 0, 2, 7,
+                                            rng_default)
     # level two:
-    thirty += _generate_dressing_problems(4, 7, 2, 4, 1, 3, 1, 3, 0, 2, 7,
-                                          rng_default)
+    problems += _generate_dressing_problems(4, 7, 2, 4, 1, 3, 1, 3, 0, 2, 7,
+                                            rng_default)
     # level three:
-    thirty += _generate_dressing_problems(7, 11, 3, 6, 2, 4, 2, 4, 1, 4, 8,
-                                          rng_default)
+    problems += _generate_dressing_problems(7, 11, 3, 6, 2, 4, 2, 4, 1, 4, 8,
+                                            rng_default)
     # level four:
-    thirty += _generate_dressing_problems(10, 14, 4, 7, 3, 5, 3, 5, 2, 5, 8,
-                                          rng_default)
-    assert len(thirty) == 30
+    problems += _generate_dressing_problems(10, 14, 4, 7, 3, 5, 3, 5, 2, 5, 8,
+                                            rng_default)
+    assert len(problems) == 30
 
     # writing the 30 questions:
     loc = os.path.dirname(os.path.realpath(__file__))
-    for q_num, q in enumerate(thirty):
+    for q_num, q in enumerate(problems):
         q_num += 1
         file_path = loc + f'/task{str(q_num).zfill(2)}.pddl'
         if not os.path.exists(file_path):
